refactor: report package refactor (position)
